requirements.txt
asgiref==3.7.2
Django>=3.2
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary
PyJWT==2.8.0
pytz==2024.1
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1